Update to v082r11 release.
byuu says:

Emulates grayscale and color emphasis modes, improves sprite timing and
PPU bus fetching behavior with the PPU disabled.

+++ bsnes/nes/interface/interface.hpp
@@ -1,5 +1,5 @@
 struct Interface {
-  virtual void video_refresh(const uint32_t *data) {}
+  virtual void video_refresh(const uint16_t *data) {}
   virtual void audio_sample(int16_t sample) {}
   virtual int16_t input_poll(bool port, unsigned device, unsigned id) { return 0; }
